Major code cleanup + package organisation
Up to this point IdeaJS did not recognise that it was a .java file so no
plugins were utilized (due to me still being a beginner with java and
IdeaJS). This was fixed, making it easier to compile from the command
line. With the java plugins available, many mistakes of mine (where a simpler
solution existed) were pointed out and are now fixed.