Bug fix progress
A major problem with the game before this commit is that you could jump
repeatedly each turn. Now you can only jump once a turn, but a new bug
is now seen; you can jump ahead of time for future turns, but for each
turn only jump once. I think this is progress... ?